feat: implement unsaved changes confirmation and show main window
Added functionality to confirm unsaved changes before quitting the application. If there are no unsaved changes, the app will exit directly; otherwise, it will show the main window for user interaction. This change enhances user experience by preventing accidental data loss.
